Increase ms interval values to 32-bit from 16-bit

+++ test/unit-test/backoff_algorithm_utest.c
@@ -83,8 +83,8 @@ int suiteTearDown( int numFailures )
  */
 static void verifyContextData( BackoffAlgorithmContext_t * pContext,
                                uint32_t expectedAttemptsDone,
-                               uint16_t expectedNextJitterMax,
-                               uint16_t expectedMaxBackoff,
+                               uint32_t expectedNextJitterMax,
+                               uint32_t expectedMaxBackoff,
                                uint32_t expectedMaxAttempts )
 {
     TEST_ASSERT_EQUAL( expectedNextJitterMax, pContext->nextJitterMax );
@@ -133,9 +133,9 @@ void test_BackoffAlgorithm_InitializeParams_Sets_Jitter_Correctly( void )
 void test_BackoffAlgorithm_GetNextBackoff_Success_RandVal_Less_Than_Jitter_Max( void )
 {
     int iter = 1;
-    uint16_t expectedAttemptsDone = 0;
-    uint16_t expectedNextJitterMax = TEST_BACKOFF_BASE_VALUE;
-    uint16_t expectedNextBackoff = 0;
+    uint32_t expectedAttemptsDone = 0;
+    uint32_t expectedNextJitterMax = TEST_BACKOFF_BASE_VALUE;
+    uint32_t expectedNextBackoff = 0;
 
     for( ; iter < 2; iter++ )
     {
@@ -180,8 +180,8 @@ void test_BackoffAlgorithm_GetNextBackoff_Success_RandVal_Less_Than_Jitter_Max( 
 void test_BackoffAlgorithm_GetNextBackoff_Success_RandVal_More_Than_Jitter_Max( void )
 {
     int iter = 1;
-    uint16_t expectedAttemptsDone = 0;
-    uint16_t expectedNextJitterMax = TEST_BACKOFF_BASE_VALUE;
+    uint32_t expectedAttemptsDone = 0;
+    uint32_t expectedNextJitterMax = TEST_BACKOFF_BASE_VALUE;
 
     for( ; iter < 2; iter++ )
     {
@@ -199,7 +199,7 @@ void test_BackoffAlgorithm_GetNextBackoff_Success_RandVal_More_Than_Jitter_Max( 
         /* As the random value is greater than the jitter max value, the expected
          * next backoff value should be truncated to a value within the jitter window
          * for the retry attempt. */
-        uint16_t expectedNextBackoff = ( testRandomVal % ( retryParams.nextJitterMax + 1U ) );
+        uint32_t expectedNextBackoff = ( testRandomVal % retryParams.nextJitterMax ) + 1U;
 
         /* Call the BackoffAlgorithm_GetNextBackoff API a couple times. */
         TEST_ASSERT_EQUAL( BackoffAlgorithmSuccess,
